Organized files and split components into individual files
