+++ public/client/components/App.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import LoggingExercise from './mainViews/LoggingExercise';
-import Overview from './Overview.js';
+import LoggingExercise from './exercise/LoggingExercise';
+import Overview from './overview/Overview.js';
 import NavigationBar from './Navbar';
 
 export default class App extends React.Component {
