+++ server/route.js
@@ -1,6 +1,7 @@
 var webpackConfig = require('../webpack.config');
 var webpack = require('webpack');
 var express = require('express');
+var bodyParser = require('body-parser');
 var path = require('path');
 var app = express();
 
@@ -9,8 +10,8 @@ var app = express();
 
 
 app.use(require("webpack-hot-middleware")(compiler));
-// app.use(bodyParser.json());
-// app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
 
 app.use(require("webpack-dev-middleware")(compiler, {
     noInfo: true, publicPath: webpackConfig.output.publicPath
@@ -74,7 +75,8 @@ app.post('/api/exercises', exerciseController.newExercise);
 
 
 app.post('/submitUnits', function(req, res) {
-  console.log('POST request received on url submitUnits')
+  console.log('POST request received on url submitUnits');
+  console.log('body', req.body); //TODO: move data to DB
   res.end('');
 });
